15: Rename components, add copilot files to gitignore
