+++ .gitignore
@@ -15,5 +15,6 @@
 local.properties
 
 # Ignore Copilot migration agent files
-copilot.data.migration.agent.xml
-copilot.data.migration.agent.*
+.idea/copilot.data.migration.*.*
+.idea/Project_Default.xml
+.idea/studiobot.xml
