+++ app/src/main/java/com/example/conquest/screens/cosplay/EditPhoto.kt
@@ -1,4 +1,4 @@
-package com.example.conquest.screens
+package com.example.conquest.screens.cosplay
 
 import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -45,10 +45,10 @@ import com.example.conquest.components.MyFab
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class EditReferenceImageScreen(val photoId: Int)
+data class EditPhoto(val photoId: Int)
 
 @Composable
-fun EditReferenceImageScreen(
+fun EditPhoto(
     photoId: Int, navController: NavController, cosplayViewModel: CosplayViewModel = viewModel()
 ) {
     val context = LocalContext.current
