+++ app/src/main/java/com/example/conquest/screens/cosplay/NewTask.kt
@@ -1,73 +1,102 @@
-package com.example.conquest.screens
+package com.example.conquest.screens.cosplay
 
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.Cancel
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Snackbar
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.Switch
+import androidx.compose.material3.SwitchDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.example.conquest.CosplayViewModel
-import com.example.conquest.components.MyFab
+import com.example.conquest.components.DatePickerFieldToModal
 import com.example.conquest.components.getCurrentDate
+import com.example.conquest.data.entity.CosplayTask
 import kotlinx.coroutines.launch
 import kotlinx.serialization.Serializable
 import java.util.Date
 
 @Serializable
-data class EditTaskScreen(val taskId: Int)
+data class NewTask(val cosplayId: Int)
 
 @Composable
-fun EditTaskScreen(
-    taskId: Int, navController: NavController, cosplayViewModel: CosplayViewModel = viewModel()
+fun NewTask(
+    cosplayId: Int,
+    navController: NavController,
 ) {
-    val task by cosplayViewModel.getTaskById(taskId).collectAsState(initial = null)
-
-    var taskName by remember { mutableStateOf("") }
-    var done by remember { mutableStateOf(false) }
-    var alarm by remember { mutableStateOf(false) }
-    var notes by remember { mutableStateOf("") }
-    var date: Date? by remember { mutableStateOf(null) }
+    val cosplayViewModel: CosplayViewModel = viewModel()
     val snackbarHostState = remember { SnackbarHostState() }
     val coroutineScope = rememberCoroutineScope()
 
-    LaunchedEffect(task) {
-        task?.let {
-            taskName = it.taskName
-            done = it.done
-            alarm = it.alarm
-            notes = it.notes ?: ""
-            date = it.date ?: getCurrentDate()
-        }
-    }
+    var name by remember { mutableStateOf("") }
+    var done by remember { mutableStateOf(false) }
+    var alarm by remember { mutableStateOf(false) }
+    var date: Date? by remember { mutableStateOf(getCurrentDate()) }
 
     Box(modifier = Modifier.fillMaxSize()) {
         Column(
             modifier = Modifier
-                .fillMaxSize()
-                .padding(horizontal = 16.dp)
-                .padding(top = 12.dp),
-            verticalArrangement = Arrangement.spacedBy(16.dp)
+                .align(Alignment.TopCenter)
+                .fillMaxWidth(0.9f)
+                .padding(top = 70.dp, start = 16.dp, end = 16.dp),
+            verticalArrangement = Arrangement.spacedBy(20.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
         ) {
             Text(
-                text = "Edit Task",
-                style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold)
+                text = "Add Task",
+                style = MaterialTheme.typography.headlineMedium.copy(
+                    fontWeight = FontWeight.Bold, letterSpacing = 1.5.sp
+                ),
+                color = MaterialTheme.colorScheme.primary,
+                modifier = Modifier
+                    .padding(bottom = 8.dp, top = 8.dp)
+                    .fillMaxWidth(),
+                textAlign = TextAlign.Center
             )
+
             OutlinedTextField(
-                value = taskName,
-                onValueChange = { taskName = it },
-                label = { Text("Task Name") },
+                value = name,
+                onValueChange = { name = it },
+                label = { Text("Task Name*") },
                 modifier = Modifier.fillMaxWidth(),
-                shape = RoundedCornerShape(20.dp),
-                singleLine = true
+                singleLine = true,
+                shape = RoundedCornerShape(32.dp)
             )
+
             Text(
                 text = "Status",
                 style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold)
@@ -143,61 +172,72 @@ fun EditTaskScreen(
                     }
                 }
             }
-            com.example.conquest.components.DatePickerFieldToModal(
+
+            DatePickerFieldToModal(
                 label = "Task date*",
                 selectedDate = date,
                 onDateSelected = { date = it }
             )
-            OutlinedTextField(
-                value = notes,
-                onValueChange = { notes = it },
-                label = { Text("Notes") },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(120.dp),
-                shape = RoundedCornerShape(20.dp),
-                maxLines = 6
+        }
+
+        IconButton(
+            onClick = {
+                navController.navigate(MainCosplayScreen(uid = cosplayId, initialTab = 1)) {
+                    popUpTo(MainCosplayScreen(uid = cosplayId)) { inclusive = true }
+                }
+            },
+            modifier = Modifier
+                .align(Alignment.TopEnd)
+                .padding(top = 16.dp, end = 16.dp)
+                .statusBarsPadding()
+                .size(60.dp)
+        ) {
+            Icon(
+                imageVector = Icons.Default.Cancel,
+                contentDescription = "Cancel",
+                tint = MaterialTheme.colorScheme.onBackground
             )
         }
-        Row(
-            modifier = Modifier
-                .align(Alignment.BottomCenter)
-                .padding(bottom = 16.dp),
-            horizontalArrangement = Arrangement.spacedBy(24.dp)
+
+        Box(
+            modifier = Modifier.align(Alignment.BottomCenter)
         ) {
-            MyFab(
-                onClick = { navController.popBackStack() },
-                containerColor = MaterialTheme.colorScheme.errorContainer,
-                contentColor = MaterialTheme.colorScheme.primary,
-                icon = Icons.Default.Close,
-                contentDescription = "Discard"
-            )
-            MyFab(
+            FloatingActionButton(
                 onClick = {
-                    if (taskName.isBlank()) {
-                        coroutineScope.launch {
-                            snackbarHostState.showSnackbar("Task name cannot be empty!")
-                        }
-                        return@MyFab
-                    }
-                    task?.let {
-                        val updatedTask = it.copy(
-                            taskName = taskName,
+                    if (name.isNotBlank()) {
+                        val task = CosplayTask(
+                            id = 0,
+                            cosplayId = cosplayId,
                             done = done,
+                            taskName = name.trim(),
                             alarm = alarm,
-                            notes = notes,
-                            date = date
+                            notes = null,
+                            date = date,
                         )
-                        cosplayViewModel.updateTask(updatedTask)
+                        cosplayViewModel.insertTask(task)
+                        navController.navigate(MainCosplayScreen(uid = cosplayId, initialTab = 1)) {
+                            popUpTo(MainCosplayScreen(uid = cosplayId)) { inclusive = true }
+                        }
+                    } else {
+                        coroutineScope.launch {
+                            snackbarHostState.showSnackbar(
+                                message = "Please fill out all required fields!"
+                            )
+                        }
                     }
-                    navController.popBackStack()
                 },
                 containerColor = MaterialTheme.colorScheme.secondary,
                 contentColor = MaterialTheme.colorScheme.primary,
-                icon = Icons.Default.Check,
-                contentDescription = "Save"
-            )
+                shape = RoundedCornerShape(50),
+                modifier = Modifier
+                    .align(Alignment.BottomCenter)
+                    .padding(bottom = 16.dp)
+                    .navigationBarsPadding()
+            ) {
+                Icon(Icons.Default.Add, contentDescription = "Save")
+            }
         }
+
         SnackbarHost(
             hostState = snackbarHostState,
             modifier = Modifier
@@ -210,7 +250,6 @@ fun EditTaskScreen(
                     contentColor = MaterialTheme.colorScheme.primary,
                     shape = RoundedCornerShape(32.dp),
                 )
-            }
-        )
+            })
     }
 }
